requirements.txt
opencv-python>=4.5.0
easyocr>=1.6.0
numpy>=1.21.0
Pillow>=8.0.0
click>=8.0.0
tqdm>=4.62.0
torch>=1.9.0
torchvision>=0.10.0